Heroku prod/requirements.txt
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.1
greenlet==1.1.3
idna==3.3
numpy==1.23.2
pandas==1.4.4
psycopg2-binary==2.9.3
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.2.1
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.40
urllib3==1.26.12
zenpy==2.0.25
pyspark==3.4.0
py4j==0.10.9.7
tables==3.8.0
dask==2023.7.0
asyncio==3.4.3